refact: hero list card in separate place
